fix: prefer default links except on ancient Vim versions
Implements the suggestion posted in:

https://github.com/leafgarland/typescript-vim/pull/193#issuecomment-1751705278

Which aligns the plugin with what other popular plugins that override
bundled syntax files do. eg.

- https://github.com/plasticboy/vim-markdown (uses `hi def link` on
  version 508 and above)
- https://github.com/tpope/vim-markdown (uses `hi def link`
  unconditionally)

This stops highlighting from getting obliterated when activating a
colorscheme that uses `:hi clear` (which will reset highlights back to
defaults), such as a Base16 theme:

- https://github.com/chriskempson/base16-vim/blob/3be3cd82cd31acfcab9a41bad853d9c68d30478d/colors/base16-one-light.vim#L146

or Dracula:

- https://github.com/dracula/vim/blob/b2cc39273abbb6b38a3d173d2a5d8c2d1c79fc19/colors/dracula.vim#L19-L24

and presumably others.